+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### Added
+- `skippedIsNotIssue` option to not mark skipped tests as 'To Investigate' in ReportPortal.
 ### Fixed
 - Error for empty `restClientConfig` while using HTTP retries.